Added feature to change background color on NaN error
When the user submits an input that results in a NaN error in the calculator, the background color will now change to red to visually alert the user of the error. This feature helps to improve the user experience and make it easier for users to identify and correct errors in their input.